requirements.txt
jsonschema
json5
python-dotenv